ai-maga-bot/requirements.txt
aiogram==3.4.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pytest==8.3.3
pytest-asyncio==0.23.8

# Для OS интеграции
psutil==5.9.8

# Для Zoom интеграции
PyJWT==2.8.0

# Для работы с существующими компонентами AIMagistr
# Эти зависимости могут быть уже установлены в основном проекте

# Для RAG и Memory Palace
sentence-transformers==2.2.2
chromadb==0.4.18
numpy==1.24.3

# Для аудио конвертации (ffmpeg-python)
ffmpeg-python==0.2.0

# Для Yandex Vision
yandex-cloud-ml-sdk==0.2.0

# Для работы с изображениями
Pillow==10.0.1